requirements.txt
allure-pytest==2.15.0
requests==2.32.5
pytest==8.4.2
Faker==37.8.0

allure-python-commons~=2.15.0
